% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{country_list}
\alias{country_list}
\title{Get basic countries data}
\usage{
country_list(cache = TRUE)
}
\value{
data.frame of countries with the following columns:

\tabular{ll}{
 \code{iso2} \tab The 2 character ISO code. \cr
 \code{iso3} \tab The 3 character ISO code. \cr
 \code{name} \tab The name of the country (english). \cr
 \code{minx} \tab Minimum X of extent of the country. \cr
 \code{miny} \tab Minimum Y of extent of the country. \cr
 \code{maxx} \tab Maximum X of extent of the country. \cr
 \code{maxy} \tab Maximum Y of extent of the country. \cr
}
}
\description{
Returns all countries listed in DOPA.
}
\examples{
\dontrun{

# Get all countries
country_list()

}
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\url{http://dopa-services.jrc.ec.europa.eu/rest/eAdmin/get_country_list}
}

