% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{country_species_count}
\alias{country_species_count}
\title{Get species count for a country}
\usage{
country_species_count(country, rlstatus = NULL, cache = TRUE)
}
\arguments{
\item{country_id}{Character country name or numeric country code.}

\item{rlstatus}{A character vector of the IUCN Conservation Status for the
species. Default is \code{NULL} in which case species of all statuses are
requested.}

\item{cache}{Sets the cachce mode: \code{TRUE} = Use cache if available and
save to cache, \code{FALSE} = Ignore cache if available and do not save to
cache, \code{"flush"} = Ignore cache if available and save to cache.}
}
\value{
Numeric count of the species whose range intersects with the country.
}
\description{
Returns a count of the species whose range intersects with the country. This
intersection on the species range includes the Exclusive Economic Zone
boundary for the country which extends out to 200 nautical miles from the
coastline. The EEZ boundary was derived from the Flanders Marine Institute
(see \url{http://www.vliz.be/vmdcdata/marbound/} for more information).
}
\details{
For more information see \url{http://www.iucnredlist.org/technical-documents/categories-and-criteria}.
}
\examples{
\dontrun{

# Using country name
country_species_count(country="Finland")

# Using country code (156 is China)
country_species_count(country=156)

# Get only critically endangered and endangered species
country_species_count(country=156, rlstatus=c("CR", "EN"))

}
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\url{http://dopa-services.jrc.ec.europa.eu/services/especies/get_country_species_count}

\code{\link{resolve_country}} \code{\link{check_iucn_status}}
}

