% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/country.R
\name{country_species_list}
\alias{country_species_list}
\title{Get species list for a country}
\usage{
country_species_list(country, status = NULL, cache = TRUE)
}
\arguments{
\item{country}{Character country name or numeric country code.}

\item{status}{A character vector of the IUCN Conservation Status for the
species. Default is \code{NULL} in which case species of all statuses are
requested.  For more information see
\href{http://www.iucnredlist.org/technical-documents/categories-and-criteria}{here}.}

\item{cache}{Sets the cache mode: \code{TRUE} = Use cache if available and
save to cache, \code{FALSE} = Ignore cache if available and do not save to
cache, \code{"flush"} = Ignore cache if available and save to cache.}
}
\value{
A data.frame of species whose range intersects with the country. Each
species has the following information associated to it:

\tabular{ll}{
 \code{iucn_species_id} \tab IUCN Species Identifier from the Red List
 website. To find the IUCN Species Identifier, search for a species and then
 look for the number in the URL, e.g. Orang Utan is 17975 \cr
 \code{taxon} \tab The taxonomic name for the species \cr
 \code{kingdom} \tab The kingdom for the species \cr
 \code{phylum} \tab The phylum for the species \cr
 \code{class} \tab The class for the species \cr
 \code{order} \tab The order for the species \cr
 \code{family} \tab The family for the species \cr
 \code{status} \tab The IUCN Conservation Status for the species according to
 the following values: CR-Critically Endangered, EN-Endangered,
 VU-Vulnerable, NT-Near Threatened, LC-Least Concern, EX-Extinct, EW-Extinct
 in the Wild, DD-Data Deficient. For more information see
 \href{http://goo.gl/wfnuHl}{here} \cr
 \code{commonname} \tab No description \cr
 \code{language} \tab No description \cr
 \code{country_id} \tab Country ISO code \cr
 \code{country_name} \tab Country name \cr
}
}
\description{
Returns a list of the species whose range intersects with the country. This
intersection on the species range includes the Exclusive Economic Zone
boundary for the country which extends out to 200 nautical miles from the
coastline. The EEZ boundary was derived from the Flanders Marine Institute
(see \href{http://www.vliz.be/vmdcdata/marbound/}{here} for more information).
}
\examples{
\dontrun{
# Get species for New Zealand
kiwi.species <- country_species_list(country="New Zealand")

# Get only endangered species for New Zealand
endangered.kiwi.species <- country_species_list(country="New Zealand",
                                                rlstatus=c("CR", "EN", "VU"))

}
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\url{http://dopa-services.jrc.ec.europa.eu/services/ibex/especies/get_country_species_list}

\code{\link{resolve_country}} \code{\link{check_iucn_status}}
}

