% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.R
\name{country_stats}
\alias{country_stats}
\title{Get country statistics}
\usage{
country_stats(country, cache = FALSE)
}
\arguments{
\item{country}{Character country name or numeric country code.}

\item{cache}{Sets the cache mode: \code{TRUE} = Use cache if available and 
save to cache, \code{FALSE} = Ignore cache if available and do not save to 
cache, \code{"flush"} = Ignore cache if available and save to cache.}
}
\value{
A data.frame of country PA statistics. Each 
PA has the following information associated to it:

\tabular{ll}{
 \code{category} \tab IThe category for the iucn protection level \cr
 \code{area_protected} \tab The protected area of the country \cr
 \code{area_protected_perc} \tab The protected area percentage of the 
 country \cr
 \code{area_total} \tab The total area of the country \cr
 \code{countryiso} \tab The ISO numeric code of the country \cr
 \code{name} \tab The name of the country \cr
 \code{iucn_cat} \tab The numeric id for the iucn protection level \cr
}
}
\description{
Returns the country core statistics associated to the amount of PAs in 
different IUCN categories.
}
\details{
If a country has no area designated to a particulart IUCN category, 
The DOPA API will leave the category out of the response. 
\code{country_stats()} returns all categories, populationg 
non-existant with NAs.
}
\examples{
\dontrun{
# Get PA statistics for Sweden
country_stats(country="Sweden")
}
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\url{http://dopa-services.jrc.ec.europa.eu/services/ibex/ehabitat/get_country_stats_all}

\code{\link{resolve_country}}
}

