% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_iucn_pa_categories}
\alias{get_iucn_pa_categories}
\title{Get IUCN protected area categories}
\usage{
get_iucn_pa_categories()
}
\value{
A dataframe of IUCN protected area categories
}
\description{
Get IUCN protected area categories
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\keyword{internal}

