% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/country.R
\name{pa_country_stats}
\alias{pa_country_stats}
\title{Get PA statistics for a given country.}
\usage{
pa_country_stats(country, cache = FALSE)
}
\arguments{
\item{country}{Character country name or numeric country code.}

\item{cache}{Sets the cache mode: \code{TRUE} = Use cache if available and
save to cache, \code{FALSE} = Ignore cache if available and do not save to
cache, \code{"flush"} = Ignore cache if available and save to cache.}
}
\value{
A data.frame of PAs for a given country. Each
PA has the following information associated to it:

\tabular{ll}{
 \code{wdpaid} \tab The ID of the protected area, e.g. taken from WCMC
 ProtectedPlanet.org website \cr
 \code{iucn_cat} \tab The IUCN protection level category \cr
 \code{name} \tab Name of the site \cr
 \code{extent} \tab The geographic extent as xmin,ymin,xmax,ymax string \cr
 \code{hriwdpaid} \tab The id of the PA corresponding to a computed HRI
 similarity map \cr
 \code{hri} \tab The habitat replaceability index value \cr
 \code{area} \tab The reported area of the site \cr
 \code{sri} \tab The average species irreplacibility score for mammals, birds
 and amphibians \cr
 \code{pi} \tab The population pressure indeces \cr
 \code{ap} \tab The reported area of the site \cr
 \code{wdpa_wkt} \tab A simplified WKT representation of the park boundary \cr
 \code{gis_area} \tab The area of the PA computed from the geometry \cr
 \code{numterrsegms} \tab The raw number of habitat segments computed from
 biophysical variables \cr
}
}
\description{
Returns the PA core statistics for a certain country generated by JRC.
}
\examples{
\dontrun{
# Get PA country statistics for Uganda
pa_country_stats(country="Uganda")
}
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\url{http://dopa-services.jrc.ec.europa.eu/services/ibex/ehabitat/get_pa_country_stats}

\code{\link{resolve_country}}
}

