% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_dopa_response}
\alias{parse_dopa_response}
\title{Coerce a DOPA response object into a dataframe}
\usage{
parse_dopa_response(x)
}
\arguments{
\item{x}{list of lists}
}
\value{
Data frame of response data
}
\description{
DOPA responses are often structured as lists of lists. Turn this into a
data structure.
}
\details{
By default, function \code{\link[httr]{content}} parses nulls in JSON into
NULLs in R, which causes problems in data.frame coercion. Therefore replace
NULLs with NA.
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\keyword{internal}

