% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resolve_country}
\alias{resolve_country}
\title{Get ISO 3166-1 country code.}
\usage{
resolve_country(country, full.name = FALSE)
}
\arguments{
\item{country}{Character country name or numeric country code.}

\item{full.name}{Logical should the full name of the country be returned?
(default: FALSE)}
}
\value{
Numeric count of the species whose range intersects with the country.
}
\description{
Country identity can be provided as a number or as country name. If the
provided number is a valid ISO 3166-1 country code it is returned directly.
}
\details{
Package \code{\link{countrycode}} is used to resolve the value of argument
\code{country} which can be either a country name (\code{country.name}) or a
ISO 3166-1 country code (\code{iso3n}).
}
\examples{
# Using country name
code <- resolve_country("Finland")

# Using country code (156 is China)
code <- resolve_country(156)
# Getting the full name
country.name <- resolve_country(156, full.name=TRUE)

# Country code can be provided as a character string as well
code <- resolve_country("156")
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\code{\link{countrycode}}
}

