% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{wktdf2sp}
\alias{wktdf2sp}
\title{Convert a data frame with a WKT column into SpatialPolygonsDataFrame.}
\usage{
wktdf2sp(x, wkt.col, p4s = "+init=epsg:4326")
}
\arguments{
\item{x}{Data frame containing all the necessary data.}

\item{wkt.col}{Character string name of the column containing the WKT data.}

\item{p4s}{Either a character string or an object of class CRS.
(default: "+init=epsg:4326")}
}
\value{
A SpatialPolygonsDataFrame.
}
\description{
Many queries to the DOPA API return a response object that has the spatial
geometry stored as a WKT MULTIPOLYGON data. This function converts the
input data frame into a SpatialPolygonsDataFrame.
}
\details{
A functional installation of \code{rgeos} is needed.
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}

