% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{check_iucn_status}
\alias{check_iucn_status}
\title{Check whether inputs are valid IUCN status codes.}
\usage{
check_iucn_status(statuses)
}
\arguments{
\item{statuses}{A character vector of the IUCN Conservation Statuses.}
}
\value{
valid A character vector of valid items in \code{statuses}.
}
\description{
A warning is raised if a item in \code{statuses} is not found. This item
is removed from the vector. If none of the items are valid, raise an error.
}
\details{
Argument \code{statuses} is a vector containing one or several of the
following:
\tabular{rl}{
 \code{"CR"} \tab Critically Endangered \cr
 \code{"EN"} \tab Endangered \cr
 \code{"VU"} \tab Vulnerable \cr
 \code{"NT"} \tab Near Threatened \cr
 \code{"LC"} \tab Least Concern \cr
 \code{"EX"} \tab Extinct \cr
 \code{"EW"} \tab Extinct in the Wild \cr
 \code{"DD"} \tab Data Deficient \cr
}
}
\examples{
# Using country name
code <- resolve_country("Finland")

# Using country code (156 is China)
code <- resolve_country(156)

# Country code can be provided as a character string as well
code <- resolve_country("156")
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}

