% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/country.R
\name{country_list}
\alias{country_list}
\title{Get basic countries data}
\usage{
country_list(cache = TRUE)
}
\arguments{
\item{cache}{Sets the cachce mode: \code{TRUE} = Use cache if available and
save to cache, \code{FALSE} = Ignore cache if available and do not save to
cache, \code{"flush"} = Ignore cache if available and save to cache.}
}
\value{
data.frame of countries with the following columns:

\tabular{ll}{
 \code{iso2} \tab The 2 character ISO code. \cr
 \code{iso3} \tab The 3 character ISO code. \cr
 \code{name} \tab The name of the country (english). \cr
 \code{gec} \tab The GEC code (deprecated) used by the CIA World Factbook. \cr
 \code{minx} \tab Minimum X of extent of the country. \cr
 \code{miny} \tab Minimum Y of extent of the country. \cr
 \code{maxx} \tab Maximum X of extent of the country. \cr
 \code{maxy} \tab Maximum Y of extent of the country. \cr
}
}
\description{
Returns all countries listed in DOPA.
}
\examples{
\dontrun{

# Get all countries
country_list()

}
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\url{http://dopa-services.jrc.ec.europa.eu/services/dopa/ehabitat/get_country_list}
}

