% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{country_species_list}
\alias{country_species_list}
\title{Get species list for a country}
\usage{
country_species_list(country, rlstatus = NULL)
}
\arguments{
\item{country_id}{Character country name or numeric country code.}

\item{rlstatus}{A character vector of the IUCN Conservation Status for the
  species. Default is \code{NULL} in which case species of all statuses are
  requested.  For more information see
  \href{http://www.iucnredlist.org/technical-documents/categories-and-criteria}{here}.}
}
\value{
A list of species whose range intersects with the country. Each
species has the following information associated to it:

\tabular{rl}{
 \code{iucn_species_id} \tab IUCN Species Identifier from the Red List
 website. To find the IUCN Species Identifier, search for a species and then
 look for the number in the URL, e.g. Orang Utan is 17975 \cr
 \code{taxon} \tab The taxonomic name for the species \cr
 \code{min_presence_id} \tab The min_presence_id indicates the minimum
 species presence according to the following values: 0-Unknown, 1-Extant,
 2-Probably Extant, 3-Possibly Extant, 4-Possibly Extinct, 5-Extinct
 (post 1500), 6-Presence Uncertain. The protected area may intersect the
 species with more than one presence type, so this is an indication of the
 minimum presence. For more information see the IUCN Red List Metadata
 Document \href{http://goo.gl/OfET66}{here}. \cr
 \code{kingdom} \tab The kingdom for the species \cr
 \code{phylum} \tab The phylum for the species \cr
 \code{class} \tab The class for the species \cr
 \code{order} \tab The order for the species \cr
 \code{family} \tab The family for the species \cr
 \code{status} \tab The IUCN Conservation Status for the species according to
 the following values: CR-Critically Endangered, EN-Endangered,
 VU-Vulnerable, NT-Near Threatened, LC-Least Concern, EX-Extinct, EW-Extinct
 in the Wild, DD-Data Deficient. For more information see
 \href{http://goo.gl/wfnuHl}{here} \cr
 \code{assessed} \tab The date the conservation status of the species was
 last assessed \cr
 \code{commonname} \tab No description \cr
 \code{language} \tab No description \cr
}
}
\description{
Returns a list of the species whose range intersects with the country. This
intersection on the species range includes the Exclusive Economic Zone
boundary for the country which extends out to 200 nautical miles from the
coastline. The EEZ boundary was derived from the Flanders Marine Institute
(see \href{http://www.vliz.be/vmdcdata/marbound/}{here} for more information).
}
\examples{
\dontrun{


}
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\url{http://dopa-services.jrc.ec.europa.eu/services/especies/get_country_species_list}

\code{\link{resolve_country}} \code{\link{check_iucn_status}}
}

